Add new program
This program is a dumbed-down version of the original cmpsat. The main
difference is that it only considers one class of instances per solver.

This simplifies the input format somewhat, since the two input files are
now simply lists of running times (or e.g. "> 1000" if the solver timed
out).

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>